Add responsive NavBar
